[FIX] website_sale_resource_booking: Make sure resource booking partner is accessible
If someone introduces a different email address for the resource
booking, a partner with such email is searched and if not found,
created, which is correct in theory.

But the core method used for that goal (`find_or_create` in
`res.partner`) is not taking multi-company into account, searching
partners only matching by mail, and creating them in the default
company.

Due to this, this commits changes the resource booking partner
population code to perform the search and creation controlling the
involved company.

TT33276